sqrt

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -16,7 +16,7 @@
 
 int _sqrt_recursive(int n, int start, int end)
 {
-	long long mid, square;
+	long mid, square;
 
 	if (start > end)
 	{
